Use UTF-16 console output in Windows shell (1.4)
When `duckdb.exe` shell is used in default Windows terminal and
`odbc_scanner` extension is used to connect to Oracle DB - the unicode
output gets broken in console (for all subsequent queries), example:

```sql
SELECT 'Здравейте' AS hello;
```
```
UÄÄÄÄÄÄÄÄÄÄÄ¿
3   hello   3
3  varchar  3
AÄÄÄÄÄÄÄÄÄÄÄ'
3 ????????? 3
AÄÄÄÄÄÄÄÄÄÄÄU
```

Expected:

```
┌───────────┐
│   hello   │
│  varchar  │
├───────────┤
│ Здравейте │
└───────────┘
```

The problem is originally reported in duckdb/odbc-scanner#86 .

It appeared that, when Oracle ODBC driver is loaded it changes the
system locale, as returned by `setlocale(LC_ALL, NULL)`, from `C` to:

```
LC_COLLATE=C;LC_CTYPE=English_United States.1252;LC_MONETARY=C;LC_NUMERIC=C;LC_TIME=C
```

The original idea was, in `odbc_scanner`, to save the locale value
before loading new ODBC drivers and restore the locale after the
`odbc_connect` call returns. But it appeared that `setlocale` on
Windows is not process-wide, but CRT (MSVC C runtime library) -wide
([ref](https://learn.microsoft.com/en-us/cpp/c-runtime-library/global-state?view=msvc-170)).
And because after duckdb/odbc-scanner#87 the `odbc_scanner` uses its
own copy of C runtime lib - it cannot access/change the locale set by
Oracle driver.

For the same reason `main` branch builds of DuckDB are not affected,
but the problem is present in `v1.4-andium`. While the problem happens
only for a minor number of users and is largely non-blocking (only
Unicode data is broken, ASCII data is displayed correctly), it breaks
the UX for Oracle users. And Oracle (along with MSSQL) is the main
target for `odbc_scanner`, thus the intention is to fix this for
`v1.4-andium`.

It was found that changing the translation mode for stdout from
`_O_TEXT` to `_O_BINARY` ([ref](https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/setmode?view=msvc-170))
can be used as a workaround:

```c++
_setmode(_fileno(stdout), _O_BINARY)
```

But this call also appeared to be CRT-wide, so cannot be applied
selectively from `odbc_scanner`.

It was also observed, that historically on `v1.4-andium` the `fputs`
call is used to print unicode to console. Incoming UTF-8 text is first
converted to UTF-16 (with `MultiByteToWideChar`) and then converted back
to UTF-8 (perhaps in some cases it can be different multibyte encoding
here, not UTF-8) with `WideCharToMultiByte` before passing it to `fputs`
.

While on `main` this was changed to use `WriteConsoleW` passing it
UTF-16 directly. And it appeared that `WriteConsoleW` is not affected
by this problem.

It is understood that recent shell enhancements in `main` are not
intended for `v1.4-andium`, so this PR makes the minimal backport only
changing the part of `utf8_printf()` call to use `WriteConsoleW`
instead of the `fputs`.

Testing: with manual smoke checks I cannot see any differences in
console output for ordinary queries. Though I have limited experience
with the DuckDB shell (mostly use other clients) so can miss some
use-cases.

Fixes: duckdb/odbc-scanner#86